+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 0.17.5
 
 - fixed: tenant check for private contacts
-- fixed: default dashboard for new users
+- fixed: default dashboard and dashboard tab for new users
 - added information for ip-ranges
 
 0.17.4